# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/auth/register":{"post":{"tags":["Authentication"],"summary":"Register","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/refresh":{"post":{"tags":["Authentication"],"summary":"Refresh","operationId":"refresh_auth_refresh_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshAccessTokenRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/me":{"get":{"tags":["Users"],"summary":"Get Current User","operationId":"get_current_user_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"tags":["Users"],"summary":"Delete Current User","operationId":"delete_current_user_users_me_delete","responses":{"204":{"description":"Successful Response"}},"security":[{"OAuth2PasswordBearer":[]}]},"patch":{"tags":["Users"],"summary":"Update Current User","operationId":"update_current_user_users_me_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/users/{username}":{"get":{"tags":["Users"],"summary":"Get User","operationId":"get_user_users__username__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/problems/tag":{"post":{"tags":["Problems"],"summary":"Create Tag","operationId":"create_tag_problems_tag_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TagCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TagCreate"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/problems/":{"get":{"tags":["Problems"],"summary":"Get Problems","operationId":"get_problems_problems__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":5,"default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProblemResponse"},"title":"Response Get Problems Problems  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Problems"],"summary":"Create Problem","operationId":"create_problem_problems__post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_problem_problems__post"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Problems"],"summary":"Delete Problem","operationId":"delete_problem_problems__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"problem_id","in":"query","required":true,"schema":{"type":"string","title":"Problem Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/problems/{problem_id}":{"get":{"tags":["Problems"],"summary":"Get Problem By Id","operationId":"get_problem_by_id_problems__problem_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"problem_id","in":"path","required":true,"schema":{"type":"string","title":"Problem Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDetailResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/submissions/":{"post":{"tags":["Submissions"],"summary":"Create Submission","operationId":"create_submission_submissions__post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmissionCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmissionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Submissions"],"summary":"Get Submissions","operationId":"get_submissions_submissions__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"problem_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Problem Id"}},{"name":"username","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":5,"default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SubmissionHeaderResponse"},"title":"Response Get Submissions Submissions  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/submissions/{submission_id}":{"get":{"tags":["Submissions"],"summary":"Get Submission By Id","operationId":"get_submission_by_id_submissions__submission_id__get","parameters":[{"name":"submission_id","in":"path","required":true,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Submission Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmissionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_create_problem_problems__post":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"difficulty":{"$ref":"#/components/schemas/Difficulty"},"constraints":{"type":"string","title":"Constraints"},"tags":{"type":"string","title":"Tags"},"sample_io":{"type":"string","title":"Sample Io"},"input_desc":{"type":"string","title":"Input Desc"},"output_desc":{"type":"string","title":"Output Desc"},"explanation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Explanation"},"memory_limit_mb":{"type":"integer","title":"Memory Limit Mb"},"time_limit_sec":{"type":"integer","title":"Time Limit Sec"},"visibility":{"type":"boolean","title":"Visibility","default":false},"source":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source"},"editorial":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Editorial"},"tests_zip":{"type":"string","contentMediaType":"application/octet-stream","title":"Tests Zip"}},"type":"object","required":["id","title","description","difficulty","constraints","tags","sample_io","input_desc","output_desc","memory_limit_mb","time_limit_sec","tests_zip"],"title":"Body_create_problem_problems__post"},"Body_login_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_auth_login_post"},"Difficulty":{"type":"string","enum":["EASY","MEDIUM","HARD"],"title":"Difficulty"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Language":{"type":"string","enum":["C","CPP","PY"],"title":"Language"},"ProblemCreateResponse":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title"},"difficulty":{"$ref":"#/components/schemas/Difficulty"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags"},"testcases":{"type":"integer","title":"Testcases"}},"type":"object","required":["id","title","difficulty","tags","testcases"],"title":"ProblemCreateResponse"},"ProblemDetailResponse":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title"},"difficulty":{"$ref":"#/components/schemas/Difficulty"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags"},"accepted_submissions":{"type":"integer","title":"Accepted Submissions"},"description":{"type":"string","title":"Description"},"constraints":{"items":{"type":"string"},"type":"array","title":"Constraints"},"input_desc":{"type":"string","title":"Input Desc"},"output_desc":{"type":"string","title":"Output Desc"},"sample_io":{"additionalProperties":{"type":"string"},"type":"object","title":"Sample Io"},"explanation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Explanation"},"memory_limit_mb":{"type":"integer","title":"Memory Limit Mb"},"time_limit_sec":{"type":"integer","title":"Time Limit Sec"},"source":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source"},"editorial":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Editorial"},"visibility":{"type":"boolean","title":"Visibility"}},"type":"object","required":["id","title","difficulty","tags","accepted_submissions","description","constraints","input_desc","output_desc","sample_io","explanation","memory_limit_mb","time_limit_sec","source","editorial","visibility"],"title":"ProblemDetailResponse"},"ProblemResponse":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title"},"difficulty":{"$ref":"#/components/schemas/Difficulty"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags"},"accepted_submissions":{"type":"integer","title":"Accepted Submissions"}},"type":"object","required":["id","title","difficulty","tags","accepted_submissions"],"title":"ProblemResponse"},"RefreshAccessTokenRequest":{"properties":{"refresh_token":{"type":"string","title":"Refresh Token"}},"type":"object","required":["refresh_token"],"title":"RefreshAccessTokenRequest"},"SubmissionCreate":{"properties":{"problem_id":{"type":"string","title":"Problem Id"},"code":{"type":"string","title":"Code"},"language":{"$ref":"#/components/schemas/Language"}},"type":"object","required":["problem_id","code","language"],"title":"SubmissionCreate"},"SubmissionHeaderResponse":{"properties":{"id":{"type":"integer","title":"Id"},"submitted_at":{"type":"string","format":"date-time","title":"Submitted At"},"problem_id":{"type":"string","title":"Problem Id"},"username":{"type":"string","title":"Username"},"verdict":{"$ref":"#/components/schemas/Verdict"},"walltime_ms":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Walltime Ms"},"memory_kb":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Memory Kb"}},"type":"object","required":["id","submitted_at","problem_id","username","verdict","walltime_ms","memory_kb"],"title":"SubmissionHeaderResponse"},"SubmissionResponse":{"properties":{"id":{"type":"integer","title":"Id"},"language":{"$ref":"#/components/schemas/Language"},"submitted_at":{"type":"string","format":"date-time","title":"Submitted At"},"problem_id":{"type":"string","title":"Problem Id"},"username":{"type":"string","title":"Username"},"verdict":{"$ref":"#/components/schemas/Verdict"},"incorrect_testcase":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Incorrect Testcase"},"output":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Output"},"walltime_ms":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Walltime Ms"},"memory_kb":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Memory Kb"},"code":{"type":"string","title":"Code"}},"type":"object","required":["id","language","submitted_at","problem_id","username","verdict","incorrect_testcase","output","walltime_ms","memory_kb","code"],"title":"SubmissionResponse"},"TagCreate":{"properties":{"name":{"type":"string","title":"Name"},"slug":{"type":"string","title":"Slug"}},"type":"object","required":["name","slug"],"title":"TagCreate"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"conf_password":{"type":"string","title":"Conf Password"}},"type":"object","required":["username","email","password","conf_password"],"title":"UserCreate"},"UserResponse":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"is_verified":{"type":"boolean","title":"Is Verified"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["username","email","is_verified","created_at"],"title":"UserResponse"},"UserUpdate":{"properties":{"username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username"},"email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Email"},"password":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Password"},"conf_password":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Conf Password"}},"type":"object","title":"UserUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"Verdict":{"type":"string","enum":["PD","AC","WA","TLE","MLE","CE","RE","SKP"],"title":"Verdict"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/auth/login"}}}}}}